+++ cloudsim-plus/src/main/java/org/cloudbus/cloudsim/cloudlets/CloudletNull.java
@@ -45,7 +45,7 @@ final class CloudletNull implements Cloudlet {
     @Override public int getPriority() {
         return 0;
     }
-    @Override public int getDeliveryTime() {
+    @Override public long getDeliveryTime() {
         return 0;
     }
     @Override public Cloudlet setDeliveryTime(int a, int b)  {return Cloudlet.NULL;}
